business logic moved out the UI
