+++ app/lib/gen/assets.gen.dart
@@ -149,6 +149,9 @@ class $AssetsImagesAssetsGen {
   /// File path: assets/images/assets/icon_leu.svg
   SvgGenImage get iconLeu => const SvgGenImage('assets/images/assets/icon_leu.svg');
 
+  /// File path: assets/images/assets/leu_steller_bg.png
+  AssetGenImage get leuStellerBg => const AssetGenImage('assets/images/assets/leu_steller_bg.png');
+
   /// File path: assets/images/assets/logo.png
   AssetGenImage get logo => const AssetGenImage('assets/images/assets/logo.png');
 
@@ -168,6 +171,7 @@ class $AssetsImagesAssetsGen {
         assetsSend,
         assetsUp,
         iconLeu,
+        leuStellerBg,
         logo,
         mosaicBackground,
         success
