--- app/lib/page/assets/qr_code_printing/qr_code_view.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class QrCodeView extends StatefulWidget {
-  const QrCodeView({super.key});
-
-  @override
-  _QrCodeViewState createState() => _QrCodeViewState();
-}
-
-class _QrCodeViewState extends State<QrCodeView> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
-  }
-}
