+++ app/lib/page/assets/receive/receive_page.dart
@@ -1,3 +1,4 @@
+import 'package:encointer_wallet/page/assets/qr_code_printing/pages/qr_code_share_or_print_view.dart';
 import 'package:flutter/material.dart';
 import 'package:focus_detector/focus_detector.dart';
 import 'package:pausable_timer/pausable_timer.dart';
@@ -5,7 +6,7 @@ import 'package:provider/provider.dart';
 import 'package:share_plus/share_plus.dart';
 
 import 'package:encointer_wallet/common/components/encointer_text_form_field.dart';
-import 'package:encointer_wallet/common/components/qr_code_view/qr_code_image_view.dart';
+
 import 'package:encointer_wallet/common/components/wake_lock_and_brightness_enhancer.dart';
 import 'package:encointer_wallet/common/theme.dart';
 import 'package:encointer_wallet/config/consts.dart';
@@ -189,7 +190,7 @@ class _ReceivePageState extends State<ReceivePage> {
                   children: [
                     // Enhance brightness for the QR-code
                     const WakeLockAndBrightnessEnhancer(brightness: 1),
-                    QrCodeImageWithButton(
+                    QrCodeShareOrPrintView(
                       qrCode: invoice.toQrPayload(),
                       shareText: dic.assets.shareInvoice,
                       printText: dic.assets.print,
